xuocore: move more data loading into xuocore
