+++ src/Managers/FontsManager.cpp
@@ -7,6 +7,7 @@
 #include "FontsManager.h"
 #include "ColorManager.h"
 #include <xuocore/uodata.h>
+#include <xuocore/text_parser.h>
 #include "../CrossUO.h"
 #include "../Application.h"
 #include "../Renderer/RenderAPI.h"
@@ -1595,7 +1596,7 @@ uint32_t CFontsManager::GetHTMLColorFromText(astr_t &str1)
 
 void CFontsManager::GetHTMLInfoFromContent(HTML_DATA_INFO &info, const astr_t &content)
 {
-    Wisp::CTextFileParser parser({}, " =", "", "\"\"");
+    TextFileParser parser({}, " =", "", "\"\"");
     auto strings = parser.GetTokens(content);
     int size = (int)strings.size();
     for (int i = 0; i < size; i += 2)
