+++ src/Wisp.h
@@ -11,7 +11,6 @@
 #include "Wisp/WispMouse.h"
 #include "Wisp/WispPacketMessage.h"
 #include "Wisp/WispPacketReader.h"
-#include "Wisp/WispTextFileParser.h"
 #include "Wisp/WispThread.h"
 #include "Wisp/WispWindow.h"
